requirements.txt
qq-botpy==1.2.1
openai==1.51.2
python-dotenv==1.0.1
aiohttp==3.10.11
pydantic==2.9.2
loguru==0.7.2 